Address issue with blank labels crashing node table.
Note, since a blank label is being used to delete nodes, I didn't want to mess with the validation yet, so this should be improved later. But for the short term, a missing label is now saved as '' and I also check for missing labels in the node table sorting so that if one got through it'll be fixed. Not ideal, but safer for short-term.