dsri override in PR#72: do not allow change to IsAdmin; resolve in future PR
The original fix, unrelated to the UI changes in PR#72, makes it so just checks to see if `?isAdmin=true` is in the URL. While this makes it possible to administrate a Digital Ocean server remotely, it also opens a huge hole in the server. The way that template files are written to the server appears to be completely insecure, which creates a potential massive security breach for hackers. While this is not perhaps that important for NetCreate as a relatively-unknown application at this time and unlikely to be a target of hackers, we should keep track of these kind of security issues in our code base with diligence.